+++ demo-shell/src/app/components/about/about.component.html
@@ -1,5 +1,12 @@
-<adf-about
-    [githubUrlCommitAlpha]="githubUrlCommitAlpha"
-    [version]="version"
-    [dependencies]="dependencies">
-</adf-about>
+<adf-about-github-link
+    [url]="url"
+    [version]="version">
+</adf-about-github-link>
+
+<adf-about-product-version>
+</adf-about-product-version>
+
+<adf-about-application-modules
+    [dependencies]="dependencies"
+    [showExtensions]="showExtensions">
+</adf-about-application-modules>
