+++ demo-shell/src/app/components/about/about.component.ts
@@ -25,9 +25,10 @@ import { name, version, commit, dependencies } from '../../../../../package.json
 })
 export class AboutComponent {
 
-    githubUrlCommitAlpha = `https://github.com/Alfresco/${name}/commits/${commit}`;
+    url = `https://github.com/Alfresco/${name}/commits/${commit}`;
     version = version;
     dependencies = dependencies;
+    showExtensions = true;
 
     constructor() {
