+++ docs/core/components/about-product-version.component.md
@@ -1,0 +1,22 @@
+---
+Title: About Product Version Component
+Added: v1.0.0
+Status: Experimental
+Last reviewed: 2019-09-09
+---
+
+# [About Product Version Component](../../../lib/core/about/about-product-version/about-product-version.component.ts "Defined in about-product-version.component.ts")
+
+Shows which version of Process Services (BPM) and Content Services (ECM) an application is running. It also shows the relevant license information, application status and Alfresco modules running in an application.
+
+## Basic Usage
+
+With default input values:
+
+```html
+<adf-about-product-version></adf-about-product-version>
+```
+
+## Details
+
+Use this component to display an overview of which version of Process Services and Content Services an application is running as well as the associated license information. It also displays application status and any modules running.
