+++ lib/core/about/about-application-modules/about-application-modules.component.scss
