+++ lib/core/about/about-github-link/about-github-link.component.html
@@ -1,0 +1,28 @@
+<div class="adf-about-container">
+    <h1>{{ application }}</h1>
+
+    <div>
+        <h3>{{ 'ABOUT.SOURCE_CODE.TITLE' | translate }}</h3>
+        <mat-card>
+            <p *ngIf="version">{{ 'ABOUT.VERSION' | translate }}: {{ version }}</p>
+
+            <div *ngIf="url">
+                <small>{{ 'ABOUT.SOURCE_CODE.DESCRIPTION' | translate }}</small>
+                <div>
+                    <a [href]="url" target="_blank">{{url}}</a>
+                </div>
+            </div>
+        </mat-card>
+    </div>
+
+    <h3>{{ 'ABOUT.SERVER_SETTINGS.TITLE' | translate }}</h3>
+    <small>{{ 'ABOUT.SERVER_SETTINGS.DESCRIPTION' | translate }}</small>
+    <mat-card>
+        <p>
+            {{ 'ABOUT.SERVER_SETTINGS.PROCESS_SERVICE_HOST' | translate: {value: bpmHost} }}
+        </p>
+        <p>
+            {{ 'ABOUT.SERVER_SETTINGS.CONTENT_SERVICE_HOST' | translate: {value: ecmHost} }}
+        </p>
+    </mat-card>
+</div>
