+++ lib/core/about/about-github-link/about-github-link.component.scss
@@ -1,0 +1,10 @@
+.adf-about-container {
+    padding: 10px;
+}
+
+.adf-table-version {
+    width: 60%;
+    border: 0;
+    border-spacing: 0;
+    text-align: center;
+}
