--- lib/core/about/about.component.html
@@ -1,121 +1,0 @@
-<div class="adf-about-container">
-    <h1>{{ application }}</h1>
-
-    <div>
-        <h3>{{ 'ABOUT.SOURCE_CODE.TITLE' | translate }}</h3>
-        <mat-card>
-            <p *ngIf="version">{{ 'ABOUT.VERSION' | translate }}: {{ version }}</p>
-
-            <div *ngIf="githubUrlCommitAlpha">
-                <small>{{ 'ABOUT.SOURCE_CODE.DESCRIPTION' | translate }}</small>
-                <div>
-                    <a [href]="githubUrlCommitAlpha" target="_blank">{{githubUrlCommitAlpha}}</a>
-                </div>
-            </div>
-        </mat-card>
-    </div>
-
-    <h3>{{ 'ABOUT.SERVER_SETTINGS.TITLE' | translate }}</h3>
-    <small>{{ 'ABOUT.SERVER_SETTINGS.DESCRIPTION' | translate }}</small>
-    <mat-card>
-        <p>
-            {{ 'ABOUT.SERVER_SETTINGS.PROCESS_SERVICE_HOST' | translate: { value: bpmHost } }}
-        </p>
-        <p>
-            {{ 'ABOUT.SERVER_SETTINGS.CONTENT_SERVICE_HOST' | translate: { value: ecmHost } }}
-        </p>
-    </mat-card>
-
-
-    <h3>{{ 'ABOUT.VERSIONS.TITLE' | translate }}</h3>
-    <div *ngIf="bpmVersion">
-        <mat-card>
-            <div *ngIf="bpmVersion">
-                <h3>{{ 'ABOUT.VERSIONS.PROCESS_SERVICE' | translate }}</h3>
-                <p>
-                    {{ 'ABOUT.VERSIONS.LABELS.EDITION' | translate }}: {{ bpmVersion.edition }}
-                </p>
-                <p>
-                    {{ 'ABOUT.VERSIONS.LABELS.VERSION' | translate }}: {{ bpmVersion.majorVersion }}.{{
-                    bpmVersion.minorVersion }}.{{ bpmVersion.revisionVersion }}
-                </p>
-            </div>
-
-            <div *ngIf="ecmVersion">
-                <h3>{{ 'ABOUT.VERSIONS.CONTENT_SERVICE' | translate }}</h3>
-                <p>
-                    {{ 'ABOUT.VERSIONS.LABELS.EDITION' | translate }}: {{ ecmVersion.edition }}
-                </p>
-                <p>
-                    {{ 'ABOUT.VERSIONS.LABELS.VERSION' | translate }}: {{ ecmVersion.version.display }}
-                </p>
-
-            </div>
-        </mat-card>
-    </div>
-
-    <div *ngIf="ecmVersion">
-        <h3>{{ 'ABOUT.VERSIONS.LABELS.LICENSE' | translate }}</h3>
-        <adf-datatable [data]="license"></adf-datatable>
-
-        <h3> {{ 'ABOUT.VERSIONS.LABELS.STATUS' | translate }}</h3>
-        <adf-datatable [data]="status"></adf-datatable>
-
-        <h3>{{ 'ABOUT.VERSIONS.LABELS.MODULES' | translate }}</h3>
-        <adf-datatable [data]="modules"></adf-datatable>
-    </div>
-
-    <h3>{{ 'ABOUT.PACKAGES.TITLE' | translate }}</h3>
-    <small>{{ 'ABOUT.PACKAGES.DESCRIPTION' | translate }}</small>
-    <adf-datatable [data]="dependencyEntries"></adf-datatable>
-
-    <div class="adf-extension-details-container" *ngIf="showExtensions">
-        <h3>{{ 'ABOUT.EXTENSIONS.TITLE' | translate }}</h3>
-        <mat-table [dataSource]="extensions$ | async">
-            <!-- $id Column -->
-            <ng-container matColumnDef="$id">
-                <mat-header-cell *matHeaderCellDef>{{ 'ABOUT.EXTENSIONS.TABLE_HEADERS.ID' | translate }}</mat-header-cell>
-                <mat-cell *matCellDef="let element">{{element.$id}}</mat-cell>
-            </ng-container>
-
-            <!-- $name Column -->
-            <ng-container matColumnDef="$name">
-                <mat-header-cell *matHeaderCellDef>{{ 'ABOUT.EXTENSIONS.TABLE_HEADERS.NAME' | translate }}</mat-header-cell>
-                <mat-cell *matCellDef="let element">{{element.$name}}</mat-cell>
-            </ng-container>
-
-            <!-- $version Column -->
-            <ng-container matColumnDef="$version">
-                <mat-header-cell *matHeaderCellDef>{{ 'ABOUT.EXTENSIONS.TABLE_HEADERS.VERSION' | translate }}</mat-header-cell>
-                <mat-cell *matCellDef="let element">{{element.$version}}</mat-cell>
-            </ng-container>
-
-            <!-- $vendor Column -->
-            <ng-container matColumnDef="$vendor">
-                <mat-header-cell *matHeaderCellDef>{{ 'ABOUT.EXTENSIONS.TABLE_HEADERS.VENDOR' | translate }}</mat-header-cell>
-                <mat-cell *matCellDef="let element">{{element.$vendor}}</mat-cell>
-            </ng-container>
-
-            <!-- $license Column -->
-            <ng-container matColumnDef="$license">
-                <mat-header-cell *matHeaderCellDef>{{ 'ABOUT.EXTENSIONS.TABLE_HEADERS.LICENSE' | translate }}</mat-header-cell>
-                <mat-cell *matCellDef="let element">{{element.$license}}</mat-cell>
-            </ng-container>
-
-            <!-- $runtime Column -->
-            <ng-container matColumnDef="$runtime">
-                <mat-header-cell *matHeaderCellDef>{{ 'ABOUT.EXTENSIONS.TABLE_HEADERS.RUNTIME' | translate }}</mat-header-cell>
-                <mat-cell *matCellDef="let element">{{element.$runtime}}</mat-cell>
-            </ng-container>
-
-            <!-- $description Column -->
-            <ng-container matColumnDef="$description">
-                <mat-header-cell *matHeaderCellDef>{{ 'ABOUT.EXTENSIONS.TABLE_HEADERS.DESCRIPTION' | translate }}</mat-header-cell>
-                <mat-cell *matCellDef="let element">{{element.$description}}</mat-cell>
-            </ng-container>
-
-            <mat-header-row *matHeaderRowDef="extensionColumns"></mat-header-row>
-            <mat-row *matRowDef="let row; columns: extensionColumns;"></mat-row>
-        </mat-table>
-    </div>
-</div>
