+++ lib/core/about/public-api.ts
@@ -15,6 +15,8 @@
  * limitations under the License.
  */
 
-export * from './about.component';
+export * from './about-github-link/about-github-link.component';
+export * from './about-product-version/about-product-version.component';
+export * from './about-application-modules/about-application-modules.component';
 
 export * from './about.module';
